Add files via upload
Nowadays, many different technologies are used to the electronic door lock using Keypad password based on AVR
microcontroller. The main processes of this project are getting user ID and passcode to check if true to open the door
be servo motor. If something not true the door will not open and the buzzer will give peep alarm.passcode could
be be chaged either by student or proffesor in admin mode.